lol-pipeline-admin/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "lol-pipeline-admin"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "lol-pipeline-common>=1.0.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "fakeredis[aioredis]>=2.2.0",
    "respx",
    "freezegun",
    "testcontainers",
    "pytest-xdist",
    "lupa",
    "pact-python>=2.0.0",
    "jsonschema>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests/unit"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes — undefined names, unused imports
    "I",    # isort — deterministic import ordering
    "B",    # flake8-bugbear — likely bugs and design issues
    "C90",  # McCabe cyclomatic complexity
    "UP",   # pyupgrade — enforce modern Python 3.12 idioms
    "N",    # pep8-naming — consistent naming conventions
    "S",    # flake8-bandit — security vulnerability scanning
    "ANN",  # flake8-annotations — require type annotations
    "SIM",  # flake8-simplify — reduce unnecessary verbosity
    "PLR",  # pylint refactoring — branch/statement/arg limits
    "RUF",  # ruff-native rules
]
ignore = [
    "ANN401",  # `Any` allowed in serialization/middleware boundary code
    "PLR2004", # magic values — HTTP status codes, stream lengths are idiomatic
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pylint]
max-branches = 12
max-statements = 50
max-args = 7
max-returns = 6

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "ANN", "SIM"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
